requirement.txt
sentence-transformers
chromadb
langchain-community 
langchain-huggingface 
langchain-chroma 